dg.fromString(dgread)	R Documentation

Convert memory buffer to dynamic group

Description

Given a character string, this function converts that into a SEXP representation representing the corresponding dynamic group from dlsh

Usage

dg.fromString(s,convert.underscore=FALSE)

Arguments

s 	 a character vector continaing the string representation of the dg

convert.underscore 	 Convert "_" in dgz dynlist names to "." in R names?

Details

A dynamic group that is converted to a string from dlsh using the dg_toString function can be loaded into R (using file or socket connections) and the buffer can be converted to an equivalent list in R.

Value

a list

Author(s)

David Sheinberg

References

DLSH documentation

See Also

Examples

[Package dgread version 1.01 Index]