dg.get(dgread)	R Documentation

Bring dg from tcl to R

Description

Dynamic groups are found and imported from the tcl to R

Usage

dg.get(groupname, convert.underscore=FALSE)

Arguments

groupname 	 a dg in the tcltk interpreter

convert.underscore 	 Convert "_" in dgz dynlist names to "." in R names?

Details

dg.get looks for an existing group in the tcl environment and if it exists, it writes it to a temp file (using dg_write) and reads it back (using dg.read). The result is an R list.

Value

a list

Author(s)

David Sheinberg

References

DLSH documentation

See Also

Examples

  dg.get("group1")

[Package dgread version 1.01 Index]