read.dgz(dgread)	R Documentation

Read DLSH .dg/.dgz data files

Description

Reads a file store from dlsh binary format into a S list.

Usage

read.dgz(file, convert.underscore=FALSE)

Arguments

file 	 a filename as a character string.

convert.underscore 	 Convert "_" in dgz dynlist names to "." in R names?

Details

Each dynlist becomes a list in the output vector. Lists of lists are converted as such. Float lists become doubles, and short and chars become ints. Strings and ints are converted as expected.

Note that temporary files for decompressing gzip'd dgz file are necessary and will be created and deleted by this process. The name of the temporary file is determined using the "tempnam" function.

Value

a list

Author(s)

David Sheinberg

References

DLSH documentation

See Also

Examples

  read.dgz("test.dgz")

[Package dgread version 1.01 Index]